Allow pass an option state_aws_region for s3 state
Continuation of the change done previously which breaks backwards
compatibility.

When deploying a project in multiple regions, this code would use the
same bucket located in eu-west-1 for all of them. This would mean that
terraform would fail accessing the bucket when using the region
eu-west-1 with an error like:

     Error loading state: AuthorizationHeaderMalformed: The authorization
     header is malformed; the region 'us-east-1' is wrong; expecting
     'eu-west-1'
     status code: 400, request id: 786C67FDB3FCDE5A, host id:
     6+ZFE4vSlyJRvr6+nzIV6lXlgTT2Y3rtGNcSOAKHgoE4+T4y4g/u9jaU7/1SlMKK81E0YvnAeas=

Fixing this issue will take time, and it requires refactoring all the
setup. Instead, I will sadly and regretfully kick the can down the road
and fix this.

The solution can be add an optional argument called:

     -var state_aws_region=...

that will define which region is the state stored.